Implement delete and update authorization using Gates

+++ app/Http/Controllers/QuestionController.php
@@ -70,6 +70,10 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        // call \Gate::define in AuthServiceProvider, and use it in controller and view
+        if(\Gate::denies('update-question', $question)){
+            abort(403, 'Access denied');
+        }       
         return view("questions.edit")->with([
             'question'=>$question
         ]);
@@ -84,6 +88,9 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, 'Access denied');
+        }   
         $question->update($request->only('title', 'body'));
         return redirect('/question')->with([
             'success'=>'Your question has been updated'
@@ -97,7 +104,10 @@ class QuestionController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Question $question)
-    {
+    {   
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, 'Access denied');
+        }   
         $question->delete();
         return redirect('/question')->with([
             'success'=>'Your question has been deleted'
